requirements.txt
absl-py==2.0.0
astunparse==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
imageio==2.32.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0rc0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.1
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
openslide-python==1.3.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
Pillow==10.0.1
piqa==1.3.2
protobuf==4.25.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0rc1
tensorflow-estimator==2.15.0rc0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0rc1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
wsitools==0.1.5
xlrd==2.0.1
